utils: new function `is_readable_file()`
A more thorough check of what constitutes as *useable* and readable
file.
